ci: account for details package now requiring at least R 4.2.0
The latest details package (0.4.0, published to CRAN on 2025-02-09)
requires R >= 4.2.0.  For jobs checking R versions before that, add a
custom URL to get an older details from MPN.

While at it, update the R 3.6.3 job to pin texPreview specifically
rather than doing a more general override of the RSPM repository.
Either approach can work [*], but the more targeted pin is nice as
long as it doesn't lead to a cascade of other necessary pins.

[*] ... assuming the goal is just to resolve a dependency failure.  If
    the goal is to have a job that checks compatibility with package
    versions from some date, then setup-r's RSPM and CRAN variables
    need to be overridden with a date-pinned URL (for an example, see
    bbr's "oldest" job).